Fix trash sort and remove file method
