+++ kDrive/UI/Controller/Favorite/FavoriteViewController.swift
@@ -61,11 +61,10 @@ class FavoriteViewController: FileListViewController {
     override func updateChild(_ file: File, at index: Int) {
         // Remove file from list if it was unfavorited
         if !file.isFavorite {
-            sortedFiles.remove(at: index)
+            let fileId = sortedFiles[index].id
             DispatchQueue.main.async { [weak self] in
                 guard let self = self else { return }
-                self.collectionView.deleteItems(at: [IndexPath(row: index, section: 0)])
-                self.showEmptyViewIfNeeded(files: self.sortedFiles)
+                self.removeFileFromList(id: fileId)
             }
             return
         }
